requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.18.54
botocore==1.21.54
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
hurry.filesize==0.9
idna==3.2
jmespath==0.10.0
kombu==5.1.0
oauthlib==3.1.1
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
